+++ Userland/Services/AudioServer/ClientConnection.h
@@ -28,7 +28,7 @@ public:
 
     void did_finish_playing_buffer(Badge<BufferQueue>, int buffer_id);
     void did_change_muted_state(Badge<Mixer>, bool muted);
-    void did_change_main_mix_volume(Badge<Mixer>, int volume);
+    void did_change_main_mix_volume(Badge<Mixer>, double volume);
 
     virtual void die() override;
 
@@ -36,7 +36,7 @@ public:
 
 private:
     virtual Messages::AudioServer::GetMainMixVolumeResponse get_main_mix_volume() override;
-    virtual void set_main_mix_volume(i32) override;
+    virtual void set_main_mix_volume(double) override;
     virtual Messages::AudioServer::EnqueueBufferResponse enqueue_buffer(Core::AnonymousBuffer const&, i32, int) override;
     virtual Messages::AudioServer::GetRemainingSamplesResponse get_remaining_samples() override;
     virtual Messages::AudioServer::GetPlayedSamplesResponse get_played_samples() override;
